Rename funsor.contract module to funsor.sum_product
