+++ funsor/__init__.py
@@ -5,8 +5,8 @@ from funsor.interpreter import reinterpret
 from funsor.terms import Funsor, Number, Variable, of_shape, to_funsor
 from funsor.torch import Function, Tensor, arange, function, torch_einsum
 
-from . import (adjoint, contract, delta, distributions, domains, einsum, gaussian, handlers, interpreter, joint,
-               minipyro, ops, terms, torch)
+from . import (adjoint, delta, distributions, domains, einsum, gaussian, handlers, interpreter, joint,
+               minipyro, ops, sum_product, terms, torch)
 
 __all__ = [
     'Domain',
@@ -19,7 +19,6 @@ __all__ = [
     'arange',
     'backward',
     'bint',
-    'contract',
     'delta',
     'distributions',
     'domains',
@@ -35,6 +34,7 @@ __all__ = [
     'ops',
     'reals',
     'reinterpret',
+    'sum_product',
     'terms',
     'to_funsor',
     'torch',
