+++ funsor/sum_product.py
